Cambio en la forma de acceso y creación

+++ src/main/java/com/gasolinera/gasolinera/Coche.java
@@ -3,9 +3,6 @@ package com.gasolinera.gasolinera;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Queue;
-
-
 
 public class Coche implements Runnable {
 
@@ -19,7 +16,8 @@ public class Coche implements Runnable {
     private int nombre;
     private Gasolinera gasolinera;
     private Surtidor surtidor;
-    private boolean haRepostado =false;
+    private boolean haRepostado = false;
+
 
     public Coche(Gasolinera gasolinera) {
         this.nombre = Coche.id;
@@ -31,10 +29,9 @@ public class Coche implements Runnable {
     public void run() {
         while (true) {
             try {
-                if(!haRepostado){
+                if (!haRepostado) {
                     repostar();
                     pagar();
-                    salir();
                 }
 
             } catch (Exception e) {
@@ -44,14 +41,20 @@ public class Coche implements Runnable {
     }
 
     private void repostar() throws InterruptedException {
+
+        int numero = (int) (Math.random() * 10 + 1);
         logger.info("El coche " + nombre + " está buscando un surtidor libre");
         surtidor = gasolinera.buscarSurtidorLibre();
-        logger.info("El coche " + nombre + " ha encontrado un surtidor libre");
-        surtidor.take();
-        logger.info("El coche " + nombre + " está repostando durante {} minutos", gasolinera.getTime() / 1000);
-        Thread.sleep(gasolinera.getTime());
-        surtidor.drop();
-        logger.info("El coche " + nombre + " ha terminado de repostar");
+        Thread.sleep(numero * 1000);
+        if (surtidor != null) {
+            logger.info("El coche " + nombre + " ha encontrado un surtidor libre");
+            surtidor.take();
+            logger.info("El coche " + nombre + " está repostando durante {} minutos", gasolinera.getTime() / 1000);
+            Thread.sleep(gasolinera.getTime());
+            surtidor.drop();
+            logger.info("El coche " + nombre + " ha terminado de repostar");
+        }
+
     }
 
 
@@ -65,25 +68,8 @@ public class Coche implements Runnable {
 
     }
 
-    private void salir() throws InterruptedException {
-
-    }
-
-/*
-    private void repostar() throws InterruptedException {
-        long time = gasolinera.getTime();
-        logger.info("{} repostando durante {}minutos", nombre, time);
-        spendTime(time);
-    }
-
- */
-
     private void spendTime(long time) throws InterruptedException {
         Thread.sleep(time * 1000);
-    }
-
-    public int getNombre() {
-        return nombre;
     }
